pyproject.toml
[project]
name = "aiblox-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "pgvector>=0.4.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
    "alembic>=1.13.3",
    "asyncpg>=0.30.0",
]
